Коммит для Некиты

+++ crm/views.py
@@ -1,10 +1,16 @@
 from django.shortcuts import render,redirect
 from django.contrib import messages
-from crm.models import Student,Course,Token,Comment
+from crm.models import Student,Course,Token,Comment,CustomUser
 from django.contrib.auth import login, authenticate, logout
-from django.contrib.auth.models import User
+from django.contrib.auth.models import Group
 from django import forms
 from django.http import HttpResponse
+from django.conf import settings
+from django.core.mail import send_mail
+import random
+from django.template.loader import render_to_string
+from django.utils.html import strip_tags
+from django.core import mail
 
 class RegisterValidation(forms.Form):
     login = forms.CharField(max_length=30)
@@ -19,6 +25,7 @@ class LoginValidation(forms.Form):
 def logout_page(request):
     logout(request)
     return redirect('/login')
+
 def login_page(request):
     if request.method == 'GET':
         return render(request, 'login.html')
@@ -49,26 +56,73 @@ def register(request):
 
         userToken = request.POST.get('inviteToken','')
         if Token.objects.filter(inviteToken = userToken).exists():
-            user = User()
-            user.username = request.POST.get('login')
-            user.email = request.POST.get('email')
-            user.set_password(request.POST.get('password'))
-            user.save()
+            dbToken = Token.objects.filter(inviteToken = userToken).first()
+            if CustomUser.objects.filter(username = request.POST.get('login')).exists():
+                messages.add_message(request, messages.ERROR, 'Пользователь с таким именем уже существует!')
+                return redirect('/register')
+            else:
+                user = CustomUser()
+                user.tg_login = request.POST.get('tg_login')
+                user.username = request.POST.get('login')
+                user.email = request.POST.get('email')
+                user.set_password(request.POST.get('password'))
+                user.save()
 
-            login(request, user)
+                if dbToken.isAdmin == True:
+                    group = Group.objects.filter(name='Admin').first()
+                    user.groups.add(group)
+                else:
+                    group = Group.objects.filter(name='Teacher').first()
+                    user.groups.add(group)
 
-            return redirect('/')
+                login(request, user)
+                return redirect('/')
 
         else:
             messages.add_message(request, messages.ERROR, 'Неверный инвайт код')
             return redirect('/register')
+
+def passReset(request):
+    if request.method == 'GET':
+       return render(request,'pass-reset.html')
+
+    if request.method == 'POST':
+        if CustomUser.objects.filter(username = request.POST.get('login')).exists():
+            users = CustomUser.objects.all()
+            for user in users:
+                if user.username == request.POST.get('login'):
+                    chars = '+-/abcdefghijklnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890'
+                    password = ''
+                    length = random.randint(10,15)
+                    for i in range(length):
+                        password += random.choice(chars)
+                    email = user.email
+                    name = user.username
+                    user.set_password(password)
+                    user.save()
+                    html_message = render_to_string('email.html',{'name':name,'password':password})
+                    plain_message = strip_tags(html_message)
+                    mail.send_mail('Восстановление пароля',plain_message,settings.EMAIL_HOST_USER, [email],html_message=html_message)
+                    return redirect('/success-send')
+
+        else:
+           messages.add_message(request, messages.ERROR, 'Пользователя с таким именем не существует!')
+           return redirect('/password-reset')
+
+def sucessRender(request):
+    return render(request,'sucess-send.html')
 
 #main work
 def index(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     students = Student.objects.all()
-    return render(request,"index.html",{"students":students})
+    isAdmin = False
+
+    if request.user.groups.filter(name='Admin').exists():
+        isAdmin = True
+
+    return render(request,"index.html",{"students":students,'isAdmin':isAdmin})
 def details(request):
     if not request.user.is_authenticated:
         return redirect('/login')
@@ -77,32 +131,39 @@ def details(request):
         id = request.GET.get('id')
         user = request.user
         student = Student.objects.get(pk = id)
-        comments = Comment.objects.all()
+        comments = Comment.objects.filter(whom_comm = id).order_by('-id')
+        isAdmin = False
+        if request.user.groups.filter(name='Admin').exists():
+            isAdmin = True
 
-        return render(request,'details.html',{"student":student,'user':user,'comments':comments})
+        return render(request,'details.html',{"student":student,'user':user,'comments':comments,'isAdmin':isAdmin})
     if request.method == 'POST':
+
         comment_text = request.POST.get('text','')
         id = request.GET.get('id')
         student = Student.objects.get(pk = id)
+
 
         if comment_text == '':
             messages.add_message(request, messages.ERROR, 'Ваш комментарий пустой!')
             return redirect('/student?id={}'.format(student.id))
 
+
+
         comment = Comment()
         comment.text = comment_text
         comment.author = request.user
         comment.whom_comm = student
-
         comment.save()
         return redirect('/student?id={}'.format(student.id))
 def add(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     if request.method == 'GET':
+        if request.user.groups.filter(name='Admin').exists():
+            courses = Course.objects.all()
+            return render(request,'add.html',{'courses':courses})
 
-        courses = Course.objects.all()
-        return render(request,'add.html',{'courses':courses})
     if request.method == "POST":
         first_name = request.POST.get("first_name", '')
         last_name = request.POST.get("last_name", '')
@@ -121,6 +182,8 @@ def add(request):
         student.room = room
         student.email = email
         student.description = description
+        student.avatar = request.FILES['avatar']
+
 
         if course_id != '':
             course = Course.objects.get(pk=course_id)
@@ -130,11 +193,14 @@ def add(request):
         student.save()
 
         return redirect('/student?id={}'.format(student.id))
+
 def courseAdd(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     if request.method == 'GET':
-        return render(request,'course-add.html')
+        if request.user.groups.filter(name='Admin').exists():
+            return render(request,'course-add.html')
+
     if request.method == 'POST':
         name = request.POST.get('name')
         teacher = request.POST.get('teacher')
@@ -147,17 +213,22 @@ def courseAdd(request):
 def courseDescript(request):
     if not request.user.is_authenticated:
         return redirect('/login')
+    isAdmin = False
+    if request.user.groups.filter(name='Admin').exists():
+        isAdmin = True
     id = request.GET.get('id')
     course = Course.objects.get(pk=id)
     students = Student.objects.filter(course = course).all()
-    return render(request,'course-descript.html',{'course':course,'students':students})
+    return render(request,'course-descript.html',{'course':course,'students':students,'isAdmin':isAdmin})
+
 def courseEdit(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     id = request.GET.get('id')
     if request.method == 'GET':
-        course = Course.objects.get(pk = id)
-        return render(request,'course-edit.html',{"course":course})
+        if request.user.groups.filter(name='Admin').exists():
+            course = Course.objects.get(pk = id)
+            return render(request,'course-edit.html',{"course":course})
     if request.method == 'POST':
         name = request.POST.get('name','')
         teacher = request.POST.get('teacher','')
@@ -170,18 +241,20 @@ def courseEdit(request):
 def deleteCourse(request):
     if not request.user.is_authenticated:
         return redirect('/login')
-    id = request.GET.get('id')
-    course = Course.objects.get(pk = id)
-    course.delete()
-    return redirect('/course-list')
+    if request.user.groups.filter(name='Admin').exists():
+        id = request.GET.get('id')
+        course = Course.objects.get(pk = id)
+        course.delete()
+        return redirect('/course-list')
 def edit(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     if request.method == 'GET':
-        courses = Course.objects.all()
-        id = request.GET.get('id')
-        student = Student.objects.get(pk = id)
-        return render(request,'change.html',{"student":student,'courses':courses})
+        if request.user.groups.filter(name='Admin').exists():
+            courses = Course.objects.all()
+            id = request.GET.get('id')
+            student = Student.objects.get(pk = id)
+            return render(request,'change.html',{"student":student,'courses':courses})
 
     if request.method == "POST":
         if not request.user.is_authenticated:
@@ -217,26 +290,34 @@ def edit(request):
 def delete(request):
     if not request.user.is_authenticated:
         return redirect('/login')
-    id = request.GET.get('id')
-    student = Student.objects.get(pk = id)
-    student.delete()
+    if request.user.groups.filter(name='Admin').exists():
+        id = request.GET.get('id')
+        student = Student.objects.get(pk = id)
+        student.delete()
 
-    return redirect('/')
+        return redirect('/')
 def deleteAll(request):
     if not request.user.is_authenticated:
         return redirect('/login')
-    students = Student.objects.all()
-    students.delete()
+    if request.user.groups.filter(name='Admin').exists():
+        students = Student.objects.all()
+        students.delete()
 
-    return redirect('/')
+        return redirect('/')
 def showCourses(request):
     if not request.user.is_authenticated:
         return redirect('/login')
+    isAdmin = False
+    if request.user.groups.filter(name='Admin').exists():
+        isAdmin = True
     courses = Course.objects.all()
-    return render(request,'course-list.html',{"courseList":courses})
+    return render(request,'course-list.html',{"courseList":courses,'isAdmin':isAdmin})
 def courseInfo(request):
     if not request.user.is_authenticated:
         return redirect('/login')
     id = request.GET.get('id')
+    isAdmin = False
     course = Course.objects.get(pk = id)
-    return render(request,'course-descript.html',{"course":course})
+    if request.user.groups.filter(name='Admin').exists():
+        isAdmin = True
+    return render(request,'course-descript.html',{"course":course,'isAdmin':isAdmin})
